Fix compilation with flags - Makefile: use -Wall -Wextra -Werror -g -std=c++17 - Remove unused private fields (_type, _line, _column, _context) from ConfigError - Use the client reference in closeCgiPipes instead of repeated _clients.at() lookups - Fix signed/unsigned comparison warnings in ErrorPageDirective and KeepaliveTimeout - Add <filesystem> include to utils.hpp for std::filesystem::remove in DELETE - example.conf: set index script.sh for /bash location; trim bash CGI test script

+++ src/configparser/directives_functions/ErrorPageDirective.cpp
@@ -41,7 +41,7 @@ void	ConfigFile::processErrorPages(Directive* directive, std::unordered_map<int,
 	std::string	URI;
 	bool		isRedirect = false;
 	int			redirectCode = -1;
-	int			numErrorCodes;
+	size_t		numErrorCodes;
 
 	numErrorCodes = directive->getParameters().size() - 1;
 	URI = directive->getParameter(numErrorCodes);
